requirements.txt
numpy==1.24.4
httpx==0.27.0
requests==2.31.0
paddlepaddle==3.0.0
streamlit-modal==0.1.2
SpeechRecognition==3.14.3
bs4==0.0.2
markdown==3.8
streamlit_modal==0.1.2